+++ src/components/DatasetEditor/index.js
@@ -7,12 +7,10 @@ import Header from "../Header"
 import EditableTitleText from "./EditableTitleText.js"
 import SamplesView from "../SamplesView"
 import SetupPage from "../SetupPage"
-import EditSampleDialog from "../EditSampleDialog"
 import useElectron from "../../hooks/use-electron"
 import useTimeToCompleteSample from "../../hooks/use-time-to-complete-sample.js"
 import TextField from "@material-ui/core/TextField"
 import { useToasts } from "../Toasts"
-import { setIn, without } from "seamless-immutable"
 import useEventCallback from "use-event-callback"
 import usePosthog from "../../hooks/use-posthog"
 import classnames from "classnames"
@@ -21,6 +19,8 @@ import useIsLabelOnlyMode from "../../hooks/use-is-label-only-mode"
 import { HotKeys } from "react-hotkeys"
 import { useHotkeyStorage } from "../HotkeyStorage"
 import useInterface from "../../hooks/use-interface"
+import useSummary from "../../hooks/use-summary"
+import useRemoveSamples from "../../hooks/use-remove-samples"
 
 import "brace/mode/javascript"
 import "brace/theme/github"
@@ -44,13 +44,9 @@ const useStyles = makeStyles({
 })
 
 export default ({
-  file,
-  datasetName = "Universal Data Tool",
-  dataset,
   content,
   inSession,
   url,
-  fileName = "unnamed",
   onChangeDataset = () => null,
   onChangeFile,
   onFileDrop,
@@ -61,14 +57,14 @@ export default ({
   selectedBrush,
 }) => {
   const labelOnlyMode = useIsLabelOnlyMode()
-  var [valueDisplay, setValueDisplay] = useState(fileName)
   const c = useStyles()
   const { addToast } = useToasts()
   const [mode, changeMode] = useState(labelOnlyMode ? "label" : initialMode)
-  const [sampleInputEditor, changeSampleInputEditor] = useState({})
   const { ipcRenderer } = useElectron() || {}
   const posthog = usePosthog()
   const { iface } = useInterface()
+  const { summary } = useSummary()
+  const removeSamples = useRemoveSamples()
 
   const [sampleIndex, setSampleIndex] = useState(null)
 
@@ -135,10 +131,10 @@ export default ({
               <EditableTitleText
                 label="File Name"
                 onChange={(newName) => {
-                  onChangeFile(setIn(file, ["fileName"], newName))
-                  setValueDisplay(newName)
+                  // onChangeFile(setIn(file, ["fileName"], newName))
+                  // setValueDisplay(newName)
                 }}
-                value={valueDisplay || ""}
+                value={""}
               />
             )
           }
@@ -149,22 +145,14 @@ export default ({
         <div style={{ height: "100%", overflowY: "scroll" }}>
           {mode === "setup" && (
             <SetupPage
-              dataset={dataset}
               onClearLabelData={() => {
-                onChangeDataset(
-                  setIn(
-                    dataset,
-                    ["samples"],
-                    dataset.samples.map((s) => without(s, "annotation"))
-                  )
-                )
+                removeSamples(summary.samples.map((s) => s._id))
               }}
               onChange={(newDataset) => {
-                const { interface: iface } = newDataset
                 if (
-                  iface.type !== iface.type &&
+                  iface.type !== newDataset.interface.type &&
                   iface.type !== "empty" &&
-                  dataset.samples.map((s) => s.annotation).some(Boolean)
+                  summary.samples.some((s) => s.hasAnnotation)
                 ) {
                   addToast(
                     "Changing label types can cause label data issues. You must clear all label data first.",
@@ -178,8 +166,6 @@ export default ({
           )}
           {mode === "samples" && (
             <SamplesView
-              file={file}
-              dataset={dataset}
               openSampleLabelEditor={(sampleIndex) => {
                 setSampleIndex(sampleIndex)
                 posthog.capture("open_sample", {
@@ -187,30 +173,12 @@ export default ({
                 })
                 changeMode("label")
               }}
-              openSampleInputEditor={(sampleIndex) => {
-                changeSampleInputEditor({ open: true, sampleIndex })
-              }}
-              deleteSample={(sampleIndex) => {
-                const newSamples = [...dataset.samples]
-                newSamples.splice(sampleIndex, 1)
-                onChangeDataset({
-                  ...dataset,
-                  samples: newSamples,
-                })
-              }}
-              onChangeFile={(file) => {
-                onChangeFile(file)
-                setValueDisplay(file.fileName)
-              }}
-              onChangeDataset={onChangeDataset}
-              authConfig={authConfig}
               user={user}
             />
           )}
           {mode === "label" && (
             <LabelView
               selectedBrush={selectedBrush}
-              dataset={dataset}
               sampleIndex={sampleIndex}
               onChangeSampleIndex={setSampleIndex}
               sampleTimeToComplete={sampleTimeToComplete}
@@ -220,26 +188,6 @@ export default ({
             />
           )}
         </div>
-        <EditSampleDialog
-          {...sampleInputEditor}
-          sampleInput={
-            sampleInputEditor.sampleIndex !== undefined
-              ? dataset.samples[sampleInputEditor.sampleIndex]
-              : null
-          }
-          onClose={() => {
-            changeSampleInputEditor({ open: false })
-          }}
-          onChange={(newInput) => {
-            onChangeDataset(
-              setIn(
-                dataset,
-                ["samples", sampleInputEditor.sampleIndex],
-                newInput
-              )
-            )
-          }}
-        />
       </div>
     </HotKeys>
   )
