+++ src/components/ImageLandmarkAnnotation/index.js
@@ -8,13 +8,12 @@ import {
   convertFromRIARegionFmt,
   convertToRIAImageFmt,
 } from "../../utils/ria-format.js"
-import useClobberedState from "../../hooks/use-clobbered-state"
 
 const Container = styled("div")({
   "& .fullscreen": { height: "100%" },
 })
 
-const [emptyObj, emptyArr] = [{}, []]
+const emptyObj = {}
 
 const posingTools = ["create-keypoints"]
