+++ src/components/TextClassification/index.js
@@ -1,6 +1,5 @@
 import React from "react"
 import NLPAnnotator from "react-nlp-annotate"
-import useClobberedState from "../../hooks/use-clobbered-state"
 
 export const TextClassification = ({
   sample,
