+++ src/components/WebApp/index.js
@@ -1,5 +1,5 @@
 // @flow
-import React, { useState, useCallback } from "react"
+import React, { useState } from "react"
 import { HeaderContext } from "../Header"
 import StartingPage from "../StartingPage"
 import DatasetEditor from "../DatasetEditor"
@@ -7,9 +7,6 @@ import ErrorToasts from "../ErrorToasts"
 import useErrors from "../../hooks/use-errors.js"
 import LocalStorageDatasetManager from "udt-dataset-managers/dist/LocalStorageDatasetManager.js"
 import useActiveDatasetManager from "../../hooks/use-active-dataset-manager"
-import download from "in-browser-download"
-import toUDTCSV from "../../utils/to-udt-csv.js"
-import { setIn } from "seamless-immutable"
 import AppErrorBoundary from "../AppErrorBoundary"
 import useEventCallback from "use-event-callback"
 import usePreventNavigation from "../../hooks/use-prevent-navigation"
@@ -17,19 +14,10 @@ import { FileContext } from "../FileContext"
 import usePosthog from "../../hooks/use-posthog"
 import ManagePluginsDialog from "../ManagePluginsDialog"
 import usePluginProvider from "../PluginProvider"
-const randomId = () => Math.random().toString().split(".")[1]
+import download from "in-browser-download"
+import toUDTCSV from "../../utils/to-udt-csv"
 
 export default () => {
-  // const {
-  //   file,
-  //   setFile,
-  //   openFile,
-  //   openUrl,
-  //   makeSession,
-  //   recentItems,
-  //   changeRecentItems,
-  // } = useFileHandler()
-
   const [datasetManager, setDatasetManager] = useActiveDatasetManager()
 
   usePreventNavigation(Boolean(datasetManager))
@@ -54,45 +42,23 @@ export default () => {
   const onClickManagePlugins = useEventCallback(() =>
     setManagePluginsDialogOpen(true)
   )
-  const onDownload = useEventCallback((format) => {
-    // if (!file) return
-    // posthog.capture("download_file", { file_type: format })
-    // const userProvidedFileName = (file.sessionId || file.fileName).replace(
-    //   /\.udt\.(csv|json)/,
-    //   ""
-    // )
-    // const outputName = userProvidedFileName + ".udt." + format
-    // if (format === "json") {
-    //   // download(JSON.stringify(file.content), outputName)
-    // } else if (format === "csv") {
-    //   // download(toUDTCSV(file.content), outputName)
-    // }
+  const onDownload = useEventCallback(async (format) => {
+    posthog.capture("download_file", { file_type: format })
+    const ds = await datasetManager.getDataset()
+    const userProvidedFileName = (
+      datasetManager.sessionId ||
+      ds.name ||
+      "MyDataset"
+    ).replace(/\.udt\.(csv|json)/, "")
+    const outputName = userProvidedFileName + ".udt." + format
+    if (format === "json") {
+      download(JSON.stringify(ds), outputName)
+    } else if (format === "csv") {
+      download(toUDTCSV(ds), outputName)
+    }
   })
 
-  const inSession = false // file && file.mode === "server"
   const [sessionBoxOpen, changeSessionBoxOpen] = useState(false)
-
-  const onJoinSession = useCallback(
-    async (sessionName) => {
-      // await openUrl(sessionName)
-    },
-    [
-      /* openUrl */
-    ]
-  )
-
-  const onLeaveSession = useEventCallback(
-    () => {}
-    // setFile({
-    //   ...file,
-    //   mode: "local-storage",
-    //   fileName: file.fileName || `copy_of_${file.id}`,
-    // })
-  )
-
-  const onChangeDataset = useEventCallback((newDataset) => {
-    // setFile(setIn(file, ["content"], newDataset))
-  })
 
   const isWelcomePage = !datasetManager
 
@@ -101,12 +67,6 @@ export default () => {
       <FileContext.Provider value={{} /*{ file, setFile }*/}>
         <HeaderContext.Provider
           value={{
-            // title: file
-            //   ? file.mode === "local-storage"
-            //     ? file.fileName
-            //     : file.url
-            //   : "unnamed",
-            // interfaceType: file?.content?.interface?.type,
             recentItems: [],
             changeRecentItems: () => {},
             onClickTemplate: onCreateTemplate,
@@ -114,11 +74,8 @@ export default () => {
             onClickManagePlugins,
             // onOpenFile: openFile,
             onOpenRecentItem: openRecentItem,
-            inSession,
             sessionBoxOpen,
             changeSessionBoxOpen,
-            onJoinSession,
-            onLeaveSession,
             // onCreateSession: makeSession,
             fileOpen: Boolean(datasetManager),
             onDownload,
@@ -139,12 +96,7 @@ export default () => {
             ) : (
               <AppErrorBoundary>
                 <DatasetEditor
-                  // file={file}
-                  // key={file.id}
-                  // {...file}
                   selectedBrush={selectedBrush}
-                  inSession={inSession}
-                  authConfig
                   recentItems={[] /* recentItems */}
                 />
               </AppErrorBoundary>
