+++ src/hooks/use-active-dataset-manager/index.js
@@ -1,4 +1,4 @@
-import { atom, useRecoilState, useRecoilValue } from "recoil"
+import { atom, useRecoilState } from "recoil"
 import { useCallback } from "react"
 
 export const activeDatasetManagerState = atom({
